# contracts/cycle37/openapi.yaml
openapi: 3.0.3
...
info:
  title: ServiceBooking API — цикл 37 (Дома, цикл 1)
  version: "37.1.0"
  license:
    name: Proprietary
  description: |
    Конвенции проекта, действующие БЕЗ ИЗМЕНЕНИЙ:
      * camelCase; enum — строками (имена членов C#); DateTime — ISO-8601 UTC; DateOnly — "YYYY-MM-DD";
        время суток — "HH:mm";
      * ДЕНЬГИ ВЕРТИКАЛИ — ЦЕЛЫЕ РУБЛИ (integer, суффикс Rub);
      * тела осознанных 400/402/429 — ГОЛАЯ СТРОКА (text/plain) по-русски; 401/403 — пустое тело; 404 — пустое тело;
      * 451 — гейты правовых документов;
      * текст, который видит человек (статусы, причины, отказы, суммы возврата, строки суммы), собирает СЕРВЕР.

    ИСКЛЮЧЕНИЕ, КАК У ЗАКАЗОВ: все 409 маршрутов /api/stays/* — JSON с машиночитаемым code и готовым message
    (StayRefusalDto, StayGuestConflictDto, StayStaffConflictDto, StaysConflictDto).

    ГЛАВНЫЕ ПРАВИЛА, КОТОРЫЕ СХЕМА ЗАКРЕПЛЯЕТ ФОРМОЙ:
      1. Ночь D начинается в дату D; период [startDate, endDate) — endDate это дата выезда.
      2. Реквизиты для оплаты есть ТОЛЬКО в PublicStayBookingDto (по токену брони) и в PaymentDetailsDto (владельцу).
         Ни один публичный DTO дома/компании/каталога их не содержит (Т37-04).
      3. ФИО исполнителя-физлица — только в ProviderFullDto (страница брони, кабинет владельца); публично — ProviderPublicDto.
      4. Сумма возврата — «не меньше X» (StayRefundViewDto.refundAtLeastRub), считает сервер (ЮР-1).
...
servers:
  - url: http://localhost:5000
...
  - url: http://localhost:4037
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CompanyId:
      name: companyId
      in: path
      required: true
      description: Id компании «Дома» (Company.Id, Kind = Stays). Салон/магазин/чужая → 404.
      schema: { type: string, format: uuid }
    HouseId:
      name: houseId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Slug:
      name: slug
      in: path
      required: true
      description: Адрес компании (dom-routes.json → slugPattern).
      schema: { type: string, minLength: 1, maxLength: 100 }
    HouseSlug:
      name: houseSlug
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 100 }
...
    CheckInQuery:
      name: checkIn
      in: query
      required: false
      schema: { type: string, format: date }
    CheckOutQuery:
      name: checkOut
      in: query
      required: false
      schema: { type: string, format: date }
    GuestsQuery:
      name: guests
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 60, default: 1 }
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
  responses:
    PlainTextError:
      description: Осознанный отказ — голая строка по-русски.
      content:
        text/plain:
          schema: { type: string }
    TooManyRequests:
      description: Лимит частоты (политика или лимит номера) — строка по-русски.
      content:
        text/plain:
          schema: { type: string }
    Unauthorized:
      description: Нет или недействителен токен. Тело пустое.
    Forbidden:
      description: Член компании без нужного права. Тело пустое.
    NotFound:
      description: Не существует или не ваше — неотличимо. Тело пустое.
    NoContent:
      description: Успех без тела.
    StaysConflict:
      description: Конфликт в кабинете.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StaysConflictDto' }
    StayRefusal:
      description: Отказ расчёта/создания брони.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StayRefusalDto' }
...
  schemas:
...
    StaysMyRole:
      type: string
      enum: [Owner, Manager, Housekeeper, SuperAdmin]
    StaysPermission:
      type: string
      enum: [ManageCompany, ManageHouses, EditHouseContent, ViewBookings, ManageBookings, ManageBlocks, ViewSchedule, ViewCabinet]
...
    StayCancellationPolicy:
      type: string
      description: ЮР-1 — три законных шаблона (LEGAL_REVIEW_CYCLE37.md §6.2).
      enum: [Standard, Flexible, NoDeductions]
...
    HousePriceMode:
      type: string
      enum: [Constant, ByDates]
...
    HouseAmenity:
      type: string
      enum: [Wifi, Kitchen, Parking, Sauna, Bbq, WashingMachine, Tv, Fireplace, GearDryer, SkiStorage, LiftTransfer, Dishwasher, Terrace]
...
    StayProviderStatus:
      type: string
      enum: [Organization, IndividualEntrepreneur, SelfEmployed, Individual]
...
    TimeOfDay:
      type: string
      pattern: '^([01]\d|2[0-3]):[0-5]\d$'
      example: "14:00"

    # ───────────── общее ─────────────
...
    HouseAmenityDto:
      type: object
      required: [code, label]
      properties:
        code: { $ref: '#/components/schemas/HouseAmenity' }
        label: { type: string }
    GateDto:
      type: object
      required: [accepting]
      properties:
        accepting: { type: boolean }
        reasonCode: { allOf: [ { $ref: '#/components/schemas/NotAcceptingReason' } ], nullable: true }
        reasonText: { type: string, nullable: true }
    ProviderPublicDto:
      type: object
      nullable: true
      description: Публично. У SelfEmployed/Individual name, ogrn, claimsAddress = null (ЮР-3).
      required: [status, statusLabel, inn]
      properties:
        status: { $ref: '#/components/schemas/StayProviderStatus' }
        statusLabel: { type: string }
        name: { type: string, nullable: true }
        inn: { type: string }
        ogrn: { type: string, nullable: true }
        claimsAddress: { type: string, nullable: true }
    ProviderFullDto:
      type: object
      nullable: true
      required: [status, statusLabel, name, inn, claimsAddress]
      properties:
        status: { $ref: '#/components/schemas/StayProviderStatus' }
        statusLabel: { type: string }
        name: { type: string }
        inn: { type: string }
        ogrn: { type: string, nullable: true }
        claimsAddress: { type: string }
    NightPriceDto:
      type: object
      required: [date, priceRub]
      properties:
        date: { type: string, format: date }
        priceRub: { type: integer }
    StayChargeLineDto:
      type: object
      required: [kind, label, quantity, unitPriceRub, nights, amountRub, prepayEligible]
      properties:
        kind: { $ref: '#/components/schemas/StayChargeKind' }
        label: { type: string, description: подпись строки (сервер) }
        quantity: { type: integer }
        unitPriceRub: { type: integer }
        nights: { type: integer }
        amountRub: { type: integer }
        prepayEligible: { type: boolean }
...
    ProblemDto:
      type: object
      required: [code, message]
      properties:
        code: { $ref: '#/components/schemas/StayRefusalCode' }
        message: { type: string }

    # ───────────── ошибки 409 ─────────────
    StayRefusalDto:
      type: object
      required: [code, message]
      properties:
        code: { $ref: '#/components/schemas/StayRefusalCode' }
        message: { type: string }
        reasonCode: { allOf: [ { $ref: '#/components/schemas/NotAcceptingReason' } ], nullable: true }
        quote: { allOf: [ { $ref: '#/components/schemas/StayQuoteDto' } ], nullable: true, description: только у PriceChanged }
...
    StayCatalogItemDto:
      type: object
      required: [houseId, url, houseName, companyName, capacity, extraBedsMax, dogsForbidden]
      properties:
        houseId: { type: string, format: uuid }
        url: { type: string, description: относительный путь /<companySlug>/<houseSlug> }
        houseName: { type: string }
        companyName: { type: string }
        coverUrl: { type: string, nullable: true }
        coverThumbUrl: { type: string, nullable: true }
        capacity: { type: integer }
        extraBedsMax: { type: integer }
        dogsForbidden: { type: boolean }
        address: { type: string, nullable: true }
        registryNumber: { type: string, nullable: true }
        priceFromRub: { type: integer, nullable: true }
        totalRub: { type: integer, nullable: true, description: стоимость ночей за выбранные даты }
        averageNightRub: { type: integer, nullable: true }
        nights: { type: integer, nullable: true }
        availableForDates: { type: boolean, nullable: true, description: только на странице компании при выбранных датах }
    StayCatalogPage:
      type: object
      required: [items, totalCount, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/StayCatalogItemDto' }
        totalCount: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    PublicStaysCompanyDto:
      type: object
      required: [id, slug, name, available, acceptingBookings, houses, today]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        logoUrl: { type: string, nullable: true }
        description: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        available: { type: boolean }
        notAvailableText: { type: string, nullable: true }
        acceptingBookings: { type: boolean }
        notAcceptingText: { type: string, nullable: true }
        provider: { $ref: '#/components/schemas/ProviderPublicDto' }
        houses:
          type: array
          items: { $ref: '#/components/schemas/StayCatalogItemDto' }
        today: { type: string, format: date }
    HousePhotoDto:
      type: object
      required: [id, url, position]
      properties:
        id: { type: string, format: uuid }
        url: { type: string }
        thumbnailUrl: { type: string, nullable: true }
        position: { type: integer }
    ExtraBedsDto:
      type: object
      required: [enabled, max, priceRub]
      properties:
        enabled: { type: boolean }
        max: { type: integer }
        priceRub: { type: integer }
    PublicRegistryDto:
      type: object
      nullable: true
      required: [objectKind, objectKindLabel]
      properties:
        objectKind: { $ref: '#/components/schemas/HouseObjectKind' }
        objectKindLabel: { type: string }
        registryNumber: { type: string, nullable: true }
        registryUrl: { type: string, nullable: true }
    PublicStayRulesDto:
      type: object
      required: [checkInTime, checkOutTime, minNights, maxNights, horizonDays, allowGapFill, allowSameDayCheckIn, holdMinutes, prepayPercent, cancellationPolicy, cancellationSummary]
      properties:
        checkInTime: { $ref: '#/components/schemas/TimeOfDay' }
        checkOutTime: { $ref: '#/components/schemas/TimeOfDay' }
        minNights: { type: integer }
        maxNights: { type: integer }
        horizonDays: { type: integer }
        allowGapFill: { type: boolean }
        allowSameDayCheckIn: { type: boolean }
        holdMinutes: { type: integer }
        prepayPercent: { type: integer }
        cancellationPolicy: { $ref: '#/components/schemas/StayCancellationPolicy' }
        cancellationSummary: { type: string }
    PublicCompanyRefDto:
      type: object
      required: [slug, name, url]
      properties:
        slug: { type: string }
        name: { type: string }
        phone: { type: string, nullable: true }
        logoUrl: { type: string, nullable: true }
        url: { type: string }
    PublicHouseDto:
      type: object
      description: НЕ содержит реквизитов для оплаты (Т37-04).
      required: [id, slug, name, photos, capacity, extraBeds, dogsForbidden, dogFeeRub, hasCot, cotFeeRub, amenities, company, rules, acceptingBookings, available, today, timeZoneId]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        photos:
          type: array
          items: { $ref: '#/components/schemas/HousePhotoDto' }
        capacity: { type: integer }
        extraBeds: { $ref: '#/components/schemas/ExtraBedsDto' }
        dogsForbidden: { type: boolean }
        dogFeeRub: { type: integer }
        hasCot: { type: boolean }
        cotFeeRub: { type: integer }
        amenities:
          type: array
          items: { $ref: '#/components/schemas/HouseAmenityDto' }
        address: { type: string, nullable: true }
        yandexMapsUrl: { type: string, nullable: true }
        twoGisUrl: { type: string, nullable: true }
        registry: { $ref: '#/components/schemas/PublicRegistryDto' }
        company: { $ref: '#/components/schemas/PublicCompanyRefDto' }
        provider: { $ref: '#/components/schemas/ProviderPublicDto' }
        rules: { $ref: '#/components/schemas/PublicStayRulesDto' }
        priceFromRub: { type: integer, nullable: true }
        acceptingBookings: { type: boolean }
        notAcceptingText: { type: string, nullable: true }
        available: { type: boolean }
        notAvailableText: { type: string, nullable: true }
        today: { type: string, format: date }
        timeZoneId: { type: string }
    CalendarDayDto:
      type: object
      required: [date, state]
      properties:
        date: { type: string, format: date }
        state: { $ref: '#/components/schemas/CalendarDayState' }
        priceRub: { type: integer, nullable: true }
    HouseCalendarDto:
      type: object
      required: [houseId, today, from, to, minNights, maxNights, allowGapFill, allowSameDayCheckIn, lastNight, days]
      properties:
        houseId: { type: string, format: uuid }
        today: { type: string, format: date }
        from: { type: string, format: date }
        to: { type: string, format: date }
        minNights: { type: integer }
        maxNights: { type: integer }
        allowGapFill: { type: boolean }
        allowSameDayCheckIn: { type: boolean }
        lastNight: { type: string, format: date }
        days:
          type: array
          items: { $ref: '#/components/schemas/CalendarDayDto' }
    StayQuoteInput:
      type: object
      additionalProperties: false
      required: [checkIn, checkOut, adults, children, dogs, needCot]
      properties:
        checkIn: { type: string, format: date }
        checkOut: { type: string, format: date }
        adults: { type: integer, minimum: 1, maximum: 30 }
        children: { type: integer, minimum: 0, maximum: 30 }
        dogs: { type: integer, minimum: 0, maximum: 20 }
        needCot: { type: boolean }
    StayQuoteDto:
      type: object
      required: [ok, problems, nights, nightPrices, lines, extraBeds, totalRub, prepayPercent, prepayRub, dueAtCheckInRub, averageNightRub, holdMinutes, checkInTime, checkOutTime, cancellationPolicy, cancellationSummary, acceptingBookings]
      properties:
        ok: { type: boolean }
        problems:
          type: array
          items: { $ref: '#/components/schemas/ProblemDto' }
        nights: { type: integer }
        nightPrices:
          type: array
          items: { $ref: '#/components/schemas/NightPriceDto' }
        lines:
          type: array
          items: { $ref: '#/components/schemas/StayChargeLineDto' }
        extraBeds: { type: integer }
        totalRub: { type: integer }
        prepayPercent: { type: integer }
        prepayRub: { type: integer }
        dueAtCheckInRub: { type: integer }
        averageNightRub: { type: integer }
        holdMinutes: { type: integer }
        checkInTime: { $ref: '#/components/schemas/TimeOfDay' }
        checkOutTime: { $ref: '#/components/schemas/TimeOfDay' }
        cancellationPolicy: { $ref: '#/components/schemas/StayCancellationPolicy' }
        cancellationSummary: { type: string }
        acceptingBookings: { type: boolean }
        notAcceptingText: { type: string, nullable: true }
    CreateStayBookingInput:
      type: object
      additionalProperties: false
      required: [checkIn, checkOut, adults, children, dogs, needCot, guestName, notifyByMessenger, expectedTotalRub, idempotencyKey]
      properties:
        checkIn: { type: string, format: date }
        checkOut: { type: string, format: date }
        adults: { type: integer, minimum: 1, maximum: 30 }
        children: { type: integer, minimum: 0, maximum: 30 }
        dogs: { type: integer, minimum: 0, maximum: 20 }
        needCot: { type: boolean }
        guestName: { type: string, minLength: 1, maxLength: 100 }
        guestPhone: { type: string, nullable: true, description: обязателен для анонима; у вошедшего игнорируется }
        arrivalTime: { allOf: [ { $ref: '#/components/schemas/TimeOfDay' } ], nullable: true }
        comment: { type: string, maxLength: 500, nullable: true }
        notifyByMessenger: { type: boolean }
        expectedTotalRub: { type: integer, minimum: 0 }
        idempotencyKey: { type: string, format: uuid }
        captchaToken: { type: string, nullable: true }
    CreateStayBookingResponse:
      type: object
      required: [token, bookingUrl, booking]
      properties:
        token: { type: string, minLength: 43 }
        bookingUrl: { type: string }
        booking: { $ref: '#/components/schemas/PublicStayBookingDto' }

    # ───────────── бронь глазами гостя ─────────────
...
    PublicStayBookingDto:
      type: object
      required: [status, displayStatus, statusText, serverTimeUtc, house, company, checkInDate, checkOutDate, checkInTime, checkOutTime, nights, adults, children, dogs, needCot, extraBeds, guestName, guestPhoneMasked, lines, nightPrices, totalRub, prepayPercent, prepayRub, dueAtCheckInRub, paymentProofs, proofs, cancellation, notifications, availableActions]
      properties:
        status: { $ref: '#/components/schemas/StayBookingStatus' }
        displayStatus: { $ref: '#/components/schemas/StayDisplayStatus' }
        statusText: { type: string }
        serverTimeUtc: { type: string, format: date-time }
        holdExpiresAtUtc: { type: string, format: date-time, nullable: true }
        house: { $ref: '#/components/schemas/BookingHouseRefDto' }
        company: { $ref: '#/components/schemas/BookingCompanyRefDto' }
        provider: { $ref: '#/components/schemas/ProviderFullDto' }
        checkInDate: { type: string, format: date }
        checkOutDate: { type: string, format: date }
        checkInTime: { $ref: '#/components/schemas/TimeOfDay' }
        checkOutTime: { $ref: '#/components/schemas/TimeOfDay' }
        nights: { type: integer }
        adults: { type: integer }
        children: { type: integer }
        dogs: { type: integer }
        needCot: { type: boolean }
        extraBeds: { type: integer }
        arrivalTime: { allOf: [ { $ref: '#/components/schemas/TimeOfDay' } ], nullable: true }
        guestName: { type: string, nullable: true }
        guestPhoneMasked: { type: string, nullable: true }
        comment: { type: string, nullable: true }
        lines:
          type: array
          items: { $ref: '#/components/schemas/StayChargeLineDto' }
        nightPrices:
          type: array
          items: { $ref: '#/components/schemas/NightPriceDto' }
        totalRub: { type: integer }
        prepayPercent: { type: integer }
        prepayRub: { type: integer }
        dueAtCheckInRub: { type: integer }
        payment: { $ref: '#/components/schemas/PaymentInstructionsDto' }
        paymentConfirmedAtUtc: { type: string, format: date-time, nullable: true }
        paymentProofs:
          type: array
          items: { $ref: '#/components/schemas/PaymentProofDto' }
        proofs: { $ref: '#/components/schemas/ProofRulesDto' }
        cancellation: { $ref: '#/components/schemas/BookingCancellationDto' }
        statusReason: { type: string, nullable: true }
        outcomeText: { type: string, nullable: true }
        checkInInfo: { $ref: '#/components/schemas/CheckInInfoDto' }
        notifications: { $ref: '#/components/schemas/BookingNotificationsDto' }
        availableActions:
          type: array
          items: { $ref: '#/components/schemas/StayGuestAction' }
...
    PaymentDetailsDto:
      type: object
      required: [paymentDetails, paymentPurpose]
      properties:
        paymentDetails: { type: string, nullable: true, maxLength: 1000 }
        paymentPurpose: { type: string, nullable: true, maxLength: 200 }
...
    ChecklistItemDto:
      type: object
      required: [code, done, text]
      properties:
        code: { type: string, enum: [ProfileFilled, PaymentDetails, ProviderInfo, HousePublished, Plan] }
        done: { type: boolean }
        text: { type: string }
...
    StaysCompanyManageDto:
      type: object
      required: [id, name, slug, cityName, timeZoneId, isActive, showInCatalog, publicUrl, myRole, myPermissions, gate, checklist, plan]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        description: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        logoUrl: { type: string, nullable: true }
        address: { type: string, nullable: true }
        yandexMapsUrl: { type: string, nullable: true }
        twoGisUrl: { type: string, nullable: true }
        cityName: { type: string }
        timeZoneId: { type: string }
        isActive: { type: boolean }
        showInCatalog: { type: boolean }
        publicUrl: { type: string }
        myRole: { $ref: '#/components/schemas/StaysMyRole' }
        myPermissions:
          type: array
          items: { $ref: '#/components/schemas/StaysPermission' }
        settings: { allOf: [ { $ref: '#/components/schemas/StaysSettingsDto' } ], nullable: true, description: null для горничной }
        paymentDetails: { allOf: [ { $ref: '#/components/schemas/PaymentDetailsDto' } ], nullable: true, description: только при ManageCompany }
        provider: { $ref: '#/components/schemas/ProviderFullDto' }
        gate: { $ref: '#/components/schemas/GateDto' }
        checklist:
          type: array
          items: { $ref: '#/components/schemas/ChecklistItemDto' }
        plan: { $ref: '#/components/schemas/StaysPlanSummaryDto' }
        awaitingPaymentCount: { type: integer, nullable: true }
...
    HouseManageDto:
      type: object
      required: [id, slug, name, capacity, extraBeds, dogsForbidden, hasCot, amenities, priceMode, registry, registryNotice, isPublished, isArchived, position, photos, publicUrl, uncoveredDates, publishProblems, hasBookings]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        capacity: { type: integer }
        extraBeds: { $ref: '#/components/schemas/ExtraBedsDto' }
        dogsForbidden: { type: boolean }
        hasCot: { type: boolean }
        amenities:
          type: array
          items: { $ref: '#/components/schemas/HouseAmenity' }
        address: { type: string, nullable: true }
        yandexMapsUrl: { type: string, nullable: true }
        twoGisUrl: { type: string, nullable: true }
        checkInInfoText: { type: string, nullable: true }
        priceMode: { $ref: '#/components/schemas/HousePriceMode' }
        constantPriceRub: { type: integer, nullable: true }
        registry: { $ref: '#/components/schemas/HouseRegistryDto' }
        registryNotice: { $ref: '#/components/schemas/RegistryNoticeDto' }
        isPublished: { type: boolean }
        isArchived: { type: boolean }
        position: { type: integer }
        photos:
          type: array
          items: { $ref: '#/components/schemas/HousePhotoDto' }
        publicUrl: { type: string }
        uncoveredDates:
          type: array
          items: { $ref: '#/components/schemas/DateRangeDto' }
        publishProblems:
          type: array
          items: { $ref: '#/components/schemas/StaysConflictCode' }
        hasBookings: { type: boolean }

    # ───────────── шахматка, блокировки ─────────────
...
paths:
...
  /api/stays/public/amenities:
    get:
      tags: [stays-public]
      operationId: staysPublicAmenities
      summary: Справочник удобств с подписями.
      responses:
        '200':
          description: Справочник.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/HouseAmenityDto' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/stays/public/catalog:
    get:
      tags: [stays-public]
      operationId: staysPublicCatalog
      summary: Каталог домов Шерегеша с фильтрами по датам, гостям и цене.
      parameters:
        - { $ref: '#/components/parameters/CheckInQuery' }
        - { $ref: '#/components/parameters/CheckOutQuery' }
        - { $ref: '#/components/parameters/GuestsQuery' }
        - name: maxPricePerNight
          in: query
          required: false
          schema: { type: integer, minimum: 1 }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: Страница каталога.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StayCatalogPage' }
        '400': { $ref: '#/components/responses/PlainTextError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/stays/public/companies/{slug}:
    get:
      tags: [stays-public]
      operationId: staysPublicCompany
      summary: Страница компании «Дома» по адресу (ссылка, QR).
      parameters:
        - { $ref: '#/components/parameters/Slug' }
        - { $ref: '#/components/parameters/CheckInQuery' }
        - { $ref: '#/components/parameters/CheckOutQuery' }
        - { $ref: '#/components/parameters/GuestsQuery' }
      responses:
        '200':
          description: Компания и её дома.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicStaysCompanyDto' }
        '400': { $ref: '#/components/responses/PlainTextError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/stays/public/companies/{slug}/houses/{houseSlug}:
    get:
      tags: [stays-public]
      operationId: staysPublicHouse
      summary: Страница дома.
      parameters:
        - { $ref: '#/components/parameters/Slug' }
        - { $ref: '#/components/parameters/HouseSlug' }
      responses:
        '200':
          description: Дом.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicHouseDto' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/stays/public/houses/{houseId}/calendar:
    get:
      tags: [stays-public]
      operationId: staysPublicCalendar
      summary: Календарь занятости и цен дома по ночам.
      parameters:
        - { $ref: '#/components/parameters/HouseId' }
        - name: from
          in: query
          required: false
          schema: { type: string, format: date }
        - name: to
          in: query
          required: false
          schema: { type: string, format: date }
      responses:
        '200':
          description: Календарь.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HouseCalendarDto' }
        '400': { $ref: '#/components/responses/PlainTextError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/stays/public/houses/{houseId}/quote:
    post:
      tags: [stays-public]
      operationId: staysPublicQuote
      summary: Расчёт стоимости и проверка дат (ничего не резервирует, всегда 200).
      parameters:
        - { $ref: '#/components/parameters/HouseId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StayQuoteInput' }
      responses:
        '200':
          description: Расчёт.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StayQuoteDto' }
        '400': { $ref: '#/components/responses/PlainTextError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/stays/public/houses/{houseId}/bookings:
    post:
      tags: [stays-public]
      operationId: staysCreateBooking
      summary: Создать бронь (гость без входа или вошедший). Идемпотентно по idempotencyKey.
      security: [ {}, { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/HouseId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateStayBookingInput' }
      responses:
        '200':
          description: Повтор с тем же ключом — существующая бронь.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateStayBookingResponse' }
        '201':
          description: Бронь создана (Held или Confirmed при предоплате 0 %).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateStayBookingResponse' }
        '400': { $ref: '#/components/responses/PlainTextError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StayRefusal' }
        '429': { $ref: '#/components/responses/TooManyRequests' }

  # ═════════════ бронь глазами гостя ═════════════
...
  /api/stays/companies/{companyId}:
    get:
      tags: [stays-company]
      operationId: staysGetCompany
      summary: Компания в кабинете (настройки, права, гейт, чек-лист, тариф).
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/CompanyId' }
      responses:
        '200':
          description: Компания.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StaysCompanyManageDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/stays/companies/{companyId}/houses/{houseId}:
    get:
      tags: [stays-houses]
      operationId: staysGetHouse
      summary: Карточка дома в кабинете.
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/CompanyId' }
        - { $ref: '#/components/parameters/HouseId' }
      responses:
        '200':
          description: Дом.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HouseManageDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [stays-houses]
      operationId: staysDeleteHouse
      summary: Удалить дом без броней.
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/CompanyId' }
        - { $ref: '#/components/parameters/HouseId' }
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/StaysConflict' }
...